pyproject.toml
[project]
name = "bestbuyapi"
version = "2.1.0"
description = "Python Wrapper over Best Buy's REST API"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "lv10", email = "luis@lv10.me" }
]
classifiers = [
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
repository = "https://github.com/lv10/bestbuyapi"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ipdb>=0.13.13",
    "ruff>=0.9.0",
    "respx>=0.22.0",
    "pre-commit>=4.0.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: live API calls (skipped in CI by default)",
    "unit: unit tests with mocks",
]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
